fixed main price

+++ app/src/main/java/mgarzon/createbest/productcatalog/MainActivity.java
@@ -122,7 +122,7 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View view) {
                 String name = editTextName.getText().toString().trim();
-                double price = Double.parseDouble(String.valueOf(editTextPrice.getText().toString()));
+                int price = Integer.parseInt(String.valueOf(editTextPrice.getText().toString()));
                 if (!TextUtils.isEmpty(name)) {
                     updateProduct(productId, name, price);
                     b.dismiss();
@@ -139,7 +139,7 @@ public class MainActivity extends AppCompatActivity {
         });
     }
 
-    private void updateProduct(String id, String name, double price) {
+    private void updateProduct(String id, String name, int price) {
         //getting the specified product reference
         DatabaseReference dR = FirebaseDatabase.getInstance().getReference("products").child(id);
         //updating product
@@ -160,7 +160,7 @@ public class MainActivity extends AppCompatActivity {
     private void addProduct() {
         //getting the values to save
         String name = editTextName.getText().toString().trim();
-        double price = Double.parseDouble(String.valueOf(editTextPrice.getText().toString()));
+        int price = Integer.parseInt(String.valueOf(editTextPrice.getText().toString()));
 
         //checking if the value is provided
         if (!TextUtils.isEmpty(name)) {
